🔪 DELETE chatVariableResolver API proposal 💀
